Add files via upload
The library management system is for the use by librarian to keep track of books available, their authors, books borrowed, books returned, fines on each book for late returns and the borrowers in the system. The information is stored in lms database. The GUI to interact with the database and perform operations are available in Java.